Add `CycleBuilder`
