+++ crates/fj-kernel/src/builder/mod.rs
@@ -1,5 +1,9 @@
 //! API for building objects
 
+mod cycle;
 mod face;
 
-pub use self::face::{FaceBuilder, FacePolygon};
+pub use self::{
+    cycle::CycleBuilder,
+    face::{FaceBuilder, FacePolygon},
+};
